refactor(api,dispatch,platform): refactored summary and validation from projects to runs
