+++ tools/check-example-simulation-runs-succeeded
@@ -14,12 +14,13 @@ EXAMPLE_SIMULATIONS_RUNS_FILENAME = os.path.join(os.path.dirname(__file__),
                                                  'simulations', 'browse', 'example-simulations.{}.json')
 
 
-def get_validate_endpoint(biosimulations_api, validate_simulation_results_data=True):
-    """ Get the endpoint for checking the status of each simulation run
+def get_validate_endpoint(biosimulations_api, run_id, validate_simulation_results_data=True):
+    """ Get the endpoint for checking the status of a simulation run
 
     Args:
         biosimulations_api (:obj:`str`): API to use to check the status of each run
             (``dev``, ``org``, or ``local``)
+        run_id (:obj:`str`): id of the simulation run
         validate_simulation_results_data (:obj:`bool`, optional): whether to validate
             the data of the simulation results
 
@@ -31,7 +32,7 @@ def get_validate_endpoint(biosimulations_api, validate_simulation_results_data=T
     else:
         endpoint = 'https://api.biosimulations.{}'.format(biosimulations_api)
 
-    endpoint += '/projects/validate'
+    endpoint += '/runs/{}/validate'.format(run_id)
 
     if validate_simulation_results_data:
         endpoint += "?validateSimulationResultsData=true"
@@ -77,14 +78,10 @@ def get_failed_runs(examples_api, biosimulations_api, validate_simulation_result
         runs = json.load(file)
 
     # check the status of each run
-    validate_endpoint = get_validate_endpoint(biosimulations_api)
-
     failures = []
     for run in runs:
-        response = requests.post(validate_endpoint, json={
-            'id': 'new-project',
-            'simulationRun': run['id'],
-        })
+        validate_endpoint = get_validate_endpoint(biosimulations_api, run['id'])
+        response = requests.get(validate_endpoint)
         try:
             response.raise_for_status()
         except requests.exceptions.RequestException:
